+++ app/Http/Controllers/ProjetController.php
@@ -35,6 +35,10 @@ class ProjetController extends Controller
 
     public function create()
     {
+        if(!auth()->user()->hasRole('gestionnaire')){
+            return redirect()->route('programme.index');
+        }
+
         $programmes = Programme::all();
         $provinces = Province::all();
         $communes = Commune::all();
@@ -64,14 +68,16 @@ class ProjetController extends Controller
     }
 
     public function show($id){
-    // Récupérer le projet avec ses relations (programme, province, commune)
-    $projet = Projet::with(['programme', 'province', 'commune'])->findOrFail($id);
-
-    return view('projet.details', compact('projet'));
-}
+        $projet = Projet::with(['programme', 'province', 'commune'])->findOrFail($id);
+        return view('projet.details', compact('projet'));
+    }
 
     public function edit($id)
     {
+        if(!auth()->user()->hasRole('gestionnaire')){
+            return redirect()->route('programme.index');
+        }
+
         $projet = Projet::findOrFail($id);
         $programmes = Programme::all();
         $provinces = Province::all();
@@ -105,18 +111,22 @@ class ProjetController extends Controller
 
     public function destroy($id)
     {
-        try{
-        $projet = Projet::findOrFail($id);
-        $projet->delete();
-        return redirect()->route('projet.index')->with('success', 'Projet supprimé avec succès.');
-    } catch (QueryException $e) {
-        if ($e->getCode() == '23000') {
-            return redirect()->route('projet.index')
-                ->with('error', 'Impossible de supprimer ce projet : il est lié à un ou plusieurs sous-projets ou d’autres entités.');
+        if(!auth()->user()->hasRole('gestionnaire')){
+            return redirect()->route('programme.index');
         }
 
-        return redirect()->route('projet.index')
+        try{
+            $projet = Projet::findOrFail($id);
+            $projet->delete();
+            return redirect()->route('projet.index')->with('success', 'Projet supprimé avec succès.');
+        } catch (QueryException $e) {
+            if ($e->getCode() == '23000') {
+                return redirect()->route('projet.index')
+                    ->with('error', 'Impossible de supprimer ce projet : il est lié à un ou plusieurs sous-projets ou d’autres entités.');
+            }
+
+            return redirect()->route('projet.index')
             ->with('error', 'Une erreur est survenue lors de la suppression.');
-    }
+        }
     }
 }
